trunks: Update TpmUtility::SealData to accept custom auth_value
We want to be able to provide a custom authorization value
when sealing the data. This will allow to have data protected
by a secret and PCR state.

The intent is to use it in the new scheme to encrypt the
user vault keyset.

BUG=b:77799573
TEST=trunks_client --regression_test
CQ-DEPEND=CL:1416254

Change-Id: I0e7c310876b3f79f0a065a473626c83743b079b4
Reviewed-on: https://chromium-review.googlesource.com/1412007
Commit-Ready: Igor <igorcov@chromium.org>
Tested-by: Igor <igorcov@chromium.org>
Reviewed-by: Igor <igorcov@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>