vm_tools: garcon: Select vsock port with garcon
This works around a problem that surfaced due to a new kernel patch
where kernel assigned vsock ports now start at a random point in the
unsigned 32 bit integer space. gRPC uses signed 32 bit integers for
ports....so if the kernel picks one with the high bit set, we end up
with assertions in gRPC due to that. This was chosen as the solution
over patching gRPC (big patch, maintenance headaches) or patching the
kernel (the kernel isn't incorrect).

BUG=chromium:922694
TEST=started multiple containers, each chose a proper port

Change-Id: I7a81c145ca8d44ee3e14ad4db89fae726079b5f0
Reviewed-on: https://chromium-review.googlesource.com/1416414
Commit-Ready: Jeffrey Kardatzke <jkardatzke@google.com>
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>