shill: wifi: drop needless scan_timer_
This timer was only used for non-default (verbose-level) logging
purposes. It tried to measure the elapsed time for a given scan, but it
was bad at even that, and really, it's redundant -- this could just as
well be tracked via timestamps in net.log, which have at least
millisecond precision.

One problem with the existing timers: they were noisy when WiFi is
disabled at bootup, because every time a ScanRequest came in, we'd kick
off the scanning task, which would notice we're disabled, and transition
from kScanIdle to kScanIdle. But this still hit a
scan_timer_.GetElapsedTime() failure, since the timer had never gotten
started (there was no scan running).

So, rather than fixup any of the deficiencies here, just drop it.

BUG=none
TEST=watch net.log with WiFi disabled at startup -> less logspam
TEST=unit tests

Change-Id: I82f33525697797e4dbd43936301e2de8195c4fab
Reviewed-on: https://chromium-review.googlesource.com/1418982
Commit-Ready: Brian Norris <briannorris@chromium.org>
Tested-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>