common-mk: Make mojom_bindings_generator script injectable.
In mojom_bindings_generator.gni, the script path is hard coded pointing
to the /usr/src/libmojo-*/ directory, which is installed by libmojo
package.

For the libchrome/libmojo migration, this script is reused to build
libchrome/libmojo itself, so the script is not yet installed at that
timing.

This CL introduces the variable |mojom_bindings_generator|
and use it if defined by the invoker.

Also, the generated files will be compiled with libmojo, so
the static library is not needed for the case.
To support it, this CL splits the template into two,
generate_mojom_bindings_gen, and generate_mojom_bindings.
Several variables and rules are moved around along with this change.

BUG=chromium:910002
TEST=Build.

Change-Id: I457f14034f137c1934efaef9bd635d42e2c3b048
Reviewed-on: https://chromium-review.googlesource.com/1420447
Commit-Ready: Hidehiko Abe <hidehiko@chromium.org>
Tested-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Tatsuhisa Yamaguchi <yamaguchi@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>