vm_tools: Replace stub tremplin service with mock tremplin stub
This CL removes the last GRPC server started by the cicerone
fuzz testing framework. This GRPC server was a stub version of
the tremplin GRPC service that was connected to via unix sockets.
This CL changed it so that the test framework uses gmock to create
a mock of the tremplin GRPC stub and uses *ForTesting methods
to make the cicerone service use this instead of trying to
connect to Tremplin.

The effect of this CL is to reduce the time taken to run 1
iteration of fuzz testing from ~1 second to 4-20 milliseconds,
or a 50-250x speed up.

Note that, unlike the previous CL in this series
(crrev.com/c/1423880), this CL does require changes to the
fuzzer, which have been submitted in crrev.com/c/1436637.

BUG=chromium:904028
TEST=Ran unit tests, ran fuzz tests
CQ-DEPEND=CL:1436637

Change-Id: I84aafe9aa64d166fc03e9138556c00ec057dc3f4
Reviewed-on: https://chromium-review.googlesource.com/1436636
Commit-Ready: Fergus Dall <sidereal@google.com>
Tested-by: Fergus Dall <sidereal@google.com>
Reviewed-by: Stephen Barber <smbarber@chromium.org>