vm_tools: Cleanup IsContainerRunning dead code
The usage of IsContainerRunning was removed when we started controlling
lxd/lxc directly. This call is no longer used, so everything relating to
it can be cleaned up. It's safe to remove the protobufs because they
were only used by code that is getting removed in this same CL.

BUG=None
TEST=tast run vm.CrostiniStartEverything, unit tests pass

Change-Id: Id61b7da3446069232fd1efae67d25c3392cdb783
Reviewed-on: https://chromium-review.googlesource.com/1427579
Commit-Ready: Jeffrey Kardatzke <jkardatzke@google.com>
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>