diagnostics: remove redundant lookup table.
In the telem tool, we've been maintaining two lookup tables to
match strings from the command line to TelemetryItemEnums used
by libtelem. With the increasing number of TelemetryItemEnums,
having both tables is becoming burdensome to maintain. We can
avoid this by maintaining a single struct, which contains all
of the data necessary to construct both tables at runtime.

BUG=chromium:901126
TEST=manual and unit tests

Change-Id: I2c2363af12b66df4f837f570efa420eb78244e3d
Signed-off-by: Paul Moy <pmoy@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1430505
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>