intel: ipu3: Clear mPostProcessBufs if process type is not rotate
When the process type is not PROCESS_ROTATE,
the mPostProcessBufs[0] is the input buffer.
It must be released after it has been used.
If it isn't released, the mPostProcessBufs
will be continuous increased in some cases.
One case is the YUV down scaling.

BUG=b:123607428
TEST=Build and deploy package to platform and camera works

Change-Id: Id8b09fec9e7edb8bc8ca80f562438a80fd77e0d3
Signed-off-by: Yang Liang <liang.l.yang@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/1434054
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Wei Meng <wei.meng@intel.com>
Reviewed-by: Hung-yu Wu <hywu@chromium.org>
Reviewed-by: Wei Meng <wei.meng@intel.com>