vm_tools: concierge: Only log non-zero exit statuses
concierge is a bit spammy with its logs about child process exits.  We
don't need to know that a child process exited cleanly so only log a
message when the child process had a non-zero exit status or it was
killed by a signal.

BUG=none
TEST=observe /var/log/messages

Change-Id: Ia85218e40eac65ff84ae0520c4206186e8a308b5
Signed-off-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1437752
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Reviewed-by: Jeffrey Kardatzke <jkardatzke@google.com>