vm_tools: concierge: Suspend VMs when the device is suspended
In order to ensure that clock drift during a device suspend does not
trigger a softlockup in the guest VM watchdog, suspend VMs when the
device is suspended and resume them when the device resumes.

The Service class already had a power manager dbus proxy so replace it
with the PowerManagerClient, which better encapsulates the complexity of
dealing with suspend and resume attempts.

BUG=chromium:906523
TEST=start an allmodconfig build in the linux source directory and then
     run `powerd_dbus_suspend --wakeup_timeout=300` on the host.
     Observe the logs to see crosvm changing the VM run status to
     "Suspending" and then back to "Running".  Also observe that the VM
     does not crash on resume.

Change-Id: Ia9c3ef10ccf3cbd90e9bfc1f500f2854504f5ea8
Signed-off-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1437935
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Reviewed-by: Jeffrey Kardatzke <jkardatzke@google.com>