cryptohome: Update the vault keyset encryption to bind to PCR
The newly created vault keysets will be encrypted to be bound to PCR
according to the design from
https://docs.google.com/document/d/1mU0e7mLMVx91PnV5qVpmysqOfKhONTgAiBJYJVWbj04/edit

This change will make all the new keysets to become encrypted using the
new method, while existent keysets will remain the same. The migration
of old keysets into a new format is planned to be done one branch later
after this lands, to make sure no loss of data happens if the version this
lands on, gets rolled back.

So, the code from crypto.cc that encrypts using TPM is changed to use the
new SealToPcrWithAuthorization method and add a new flag (PCR_BOUND) to
mark the keyset as bound to PCR.

The decrypt code looks at the PCR_BOUND flag from keyset. If the flag is
not set, the old decryption code is used. Otherwise the new method
UnsealWithAuthorization is used.

BUG=b:77799573
TEST=unit_tests and check that decrypt works for old account and newly
     created account, while failing if PCR is extended.
CQ-DEPEND=CL:1425240

Change-Id: Ic6f82a824ba249f5ff447d7e7a36266918965cbe
Reviewed-on: https://chromium-review.googlesource.com/1436238
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Igor <igorcov@chromium.org>
Reviewed-by: Igor <igorcov@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>