arc: Rework on protocol over VSOCK.
This changes the protocol on VSOCK.
In the old protocol, only available data/fd is transferred to the other side,
and other operations are not explict, such as close / arc_bridge.sock
connection.

In the new protocol, everything is explicit.
This is preparation of supporting regular file pseudo fd passing.

BUG=b:123432308
TEST=Ran on DUT.

Change-Id: I1b9c306262ed0f355b51ab2e1103d1389ac4cb7e
Reviewed-on: https://chromium-review.googlesource.com/1445856
Commit-Ready: Hidehiko Abe <hidehiko@chromium.org>
Tested-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Yusuke Sato <yusukes@chromium.org>