hardware_verifier: Initial commit with CLI mode logic.
Initial setup for the package and implement the core logic of the
CLI mode.  The protobuf definition contains almost nothing because
some of the detail is still under discussion.

As most of the modules are not implemented at this moment, this
CL doesn't provide the entry point of the program.  The build
script is also set to support the testing only.

BUG=chromium:926825
TEST=`FEATURES=test emerge-eve chromeos-base/hardware_verifier`

Cq-Depend: chromium:1446011
Change-Id: Ia4de27b353de7acd5fa81f1390f49c8bb6117bb6
Reviewed-on: https://chromium-review.googlesource.com/1445857
Commit-Ready: Yong Hong <yhong@google.com>
Tested-by: Yong Hong <yhong@google.com>
Reviewed-by: Hong-Min Chu <hmchu@chromium.org>