power: Refactor dark resume code path in suspender
This change refactors and separates the dark resume code path in the
class that determined the CPU state based on events in powerd. A new
state is added to represent dark resume and an unsuccessful suspend
together as they respond to events identically. This new state is
separated from the device waiting for suspend delays i.e. transitioning
to suspend after a normal resume. Also, a dark resume legacy test is
removed as legacy dark resume will be completely removed in a following
patch.

BUG=chromium:927192
TEST=Run unit tests.

Change-Id: I1bab104107242918bc71de3c4223857c8dc08bf3
Reviewed-on: https://chromium-review.googlesource.com/1446434
Commit-Ready: Abhishek Bhardwaj <abhishekbh@chromium.org>
Tested-by: Abhishek Bhardwaj <abhishekbh@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>