vm_tools: Add DeleteContainer operation to tremplin
Eventually we will want to expose a way to create and delete LXD
containers in the UI. This CL adds a method to the tremplin proto that
will allow cicerone to tell it to delete an LXD container. See dependent
CL for implementation.

BUG=chromium:927531
TEST=manually tested
CQ-DEPEND=CL:1451617

Change-Id: I5ded5459dcafe213419e8f23667d024d60fb660f
Reviewed-on: https://chromium-review.googlesource.com/1449874
Commit-Ready: Fergus Dall <sidereal@google.com>
Tested-by: Fergus Dall <sidereal@google.com>
Reviewed-by: Stephen Barber <smbarber@chromium.org>