vm_tools: Add DeleteLxdContainer command to cicerone
Eventually we will want to expose a way to create and delete LXD
containers in the UI. There is currently no way to tell cicerone to
delete an LXD container. This CL adds a new DBus method to cicerone that
will allow it to do so.

BUG=chromium:927531
TEST=manually tested

Change-Id: Idc9896e6f64799f34c4303cbf88fd499f517431c
Reviewed-on: https://chromium-review.googlesource.com/1449875
Commit-Ready: Fergus Dall <sidereal@google.com>
Tested-by: Fergus Dall <sidereal@google.com>
Reviewed-by: Stephen Barber <smbarber@chromium.org>