arc: networkd: Prevent process shutdown and crash.
This patch addresses the potential for arc-networkd to get itself in
bad state, terminate or crash. The root cause is that the helper
process can still be doing work after the container goes down. This
is now possible because the daemon is always up - previously it only
ran on arc boot. The following scenarios are addressed:

1. Crash on clean up. On container down, the daemon is signaled and
cleans up IPv6 tables and routes. The container may disappear and
turn down the net namespace before arc-network gets to this. In these
cases, the nsenter calls fail and PCHECK kills the process. This patch
performs better state checking and removes some PCHECKs.

2. Self-termination by mistake. This can happen if the user logs in and
out quickly, or the container goes down before the daemon can determine
if the container is ready (i.e. container brings up arc0). EnableInbound
will repost and retry until the container is "ready" (i.e. arc0 is
brought up). If the container goes down first, this process will continue
until the max number of retries are hit at which point it quits and dies.
This (I guess) made sense in the old world when the process would be
started a new on the next log in. So this patch terminates the enable
processing if the container was downed or even comes back up so fast
that it determines a different namespace is now active.

BUG=b:123594064
BUG=b:123672966
TEST=Flashed nocturne. Verified normal behavior. Unit tests.
TEST=Verified on device, rapid login/out sequence works properly.

Change-Id: I599d0967cbd1674635536bf15bc10a9132276d7a
Reviewed-on: https://chromium-review.googlesource.com/1447915
Commit-Ready: Garrick Evans <garrick@chromium.org>
Tested-by: Garrick Evans <garrick@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>