power: Ignore zero battery charges.
Update powerd's PowerSupply class to drop battery readings
with charges of zero. It previously only dropped these when
the nominal voltage was also zero, but it looks like we're
still seeing bogus readings that can confuse users when
displayed in the UI.

Also refine the handling of zero charges when multiple
batteries are present. If just one battery reports a zero
charge, it is treated as empty and still included in
calculations, but if all batteries have zero charges,
PowerSupply discards the reading.

BUG=chromium:924869
TEST=updated unit tests

Change-Id: I3a4864f79e6fe3bb0adabcd373c0a878fa311044
Reviewed-on: https://chromium-review.googlesource.com/1449043
Commit-Ready: Dan Erat <derat@chromium.org>
Tested-by: Dan Erat <derat@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>