shill: Distinguish between an unknown failure and no failure at all
Prior to this CL, kFailureUnknown was used to represent both an unknown
failure and for no failure. This is incovenient for a number of reasons,
and led to logs indicating unknown failure when completely normal
control flow occurs.

This change adds kFailureNone to be used as the value representing no
failure. kFailureUnknown is moved to a different value such that
kFailureNone may be value 0. While this affects existing metrics, the
metrics we have on kFailureUnknown are mostly useless, as we cannot
distinguish between real failures and no failures anyways.

Note that for cellular, any disconnects from PPP still result in setting
the failure to kFailureUnknown, even if the disconnect is not the result
of any failure. Changing this behavior requires either modifying how shill
interacts with pppd through the PPP class or assuming that a PPP
disconnect after successful authentication is intentional.

A very similar CL was originally written by @semenzato many moons ago:
https://android-review.googlesource.com/c/platform/system/connectivity/shill/+/209433

BUG=chromium:596197
TEST=- Normal state changes in shill are now logged without indicating an
       unknown failure, while state changes due to actual failures do show
       a failure.
     - All unit tests are passing.

Change-Id: I76970f613ec997a6dbedf6edfaead31592ada455
Reviewed-on: https://chromium-review.googlesource.com/1452314
Commit-Ready: Alex Khouderchah <akhouderchah@chromium.org>
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>