cryptohome: Create GCache and GCache/v2 for ephemeral cryptohomes.
Cros-disks data dir preparation for DriveFS mounting relies on
<cryptohome>/GCache/v2 being accessible and writeable. This is the case
for regular cryptohomes, but not for ephemeral cryptohomes. This causes
cros-disks to fail to mount DriveFS when ephemeral mode is enabled.

Create GCache and GCache/v2 for ephemeral cryptohomes as well so DriveFS
can be mounted in ephemeral mode.

BUG=chromium:928233
TEST=unit tests, manual

Change-Id: I459336e07d880ca5edda190f3a28d21dbd2320f6
Reviewed-on: https://chromium-review.googlesource.com/1452216
Commit-Ready: Sam McNally <sammc@chromium.org>
Tested-by: Sam McNally <sammc@chromium.org>
Reviewed-by: Sergei Datsenko <dats@chromium.org>