power: powerd_suspend: remove QCA6174A from BT mode change whitelist
CL:1381326 introduced a whitelist for which BT controllers should
utilize the mode change mask, and added QCA6174A to it. This was tested
primarily on Grunt, but it caused regressions on Scarlet -- likely
because Scarlet uses a different BT wake mechanism (out-of-band GPIO)
than Grunt (in-band USB wake).

Let's drop QCA6174A IDs from the whitelist. We can try again to
re-enable this in the future.

BUG=chromium:928358, b:119700112
TEST=with or without a paired BT device, run `set_short_powerd_timeouts`
     and watch system suspend after a few seconds (Scarlet and Grunt)

Change-Id: Ib81bfd6eb32052514bad1c985774077f70cbe465
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1454323
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Reviewed-by: Dan Erat <derat@chromium.org>
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>