cryptohome: Support either-of-PCRs for TPM2 signature sealing
Update the TPM2 signature sealing backend to support restrictions
"the PCRs should be in either of the specified states", e.g.
"(PCRx==foo && PCRy==bar) OR (PCRx==bar && PCRz==baz)".

This paves the way for supporting the feature "allow login only when
specified PCRs are either in the default or in the user-specific
states" for challenge-response protected cryptohomes (a.k.a. smart
card login).

BUG=chromium:806788,b:77799573
TEST=existing unit test Tpm2RsaSignatureSecretSealingTest,
     existing tpm_live_test SignatureSealedSecretTest

Change-Id: Ieb9bcd4200cec916b26dd8d59c3656cfa4ddc7b4
Reviewed-on: https://chromium-review.googlesource.com/1455939
Commit-Ready: Maksim Ivanov <emaxx@chromium.org>
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>