system_api: Updates needed for refactoring some of Crostini app search
Changes to a proto and constants file needed for some desired
refactoring happening in CL:1447912. Before this lands
I will revert CL:1444901, so that there aren't any compile
errors from a new proto, and then reland that with the appropriate
changes later. Purpose of this refactoring is to merge
LinuxPackageInfo with LinuxPackageInfoFromApt and
InstallLinuxPackage with InstallLinuxPackageFromApt.

BUG=chromium:921429

TEST=manual

Change-Id: I147bd0469728d20c85029bcd5acd5f6d08068c18
Reviewed-on: https://chromium-review.googlesource.com/1457498
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Daniel Ng <danielng@google.com>
Reviewed-by: Jeffrey Kardatzke <jkardatzke@google.com>