cryptohome: Constant for PCR index that locks the user renamed
The constant for PCR index was initially planed to be used only for ARC
sessions. Now the scope is broader and the index is used basically to
lock the device to have access to a single user data. It is renamed to
be in line with the new meaning.

BUG=None
TEST=Code compiles and unit tests pass

Change-Id: If199ae116cff05fb9371e21c54d8ac09e7f2ba27
Reviewed-on: https://chromium-review.googlesource.com/1458576
Commit-Ready: Igor <igorcov@chromium.org>
Tested-by: Igor <igorcov@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>