vm_tools: improve set_timezone error messages
cicerone_client would print a misleading message about the number of
containers for which SetTimezone had failed, even if in many cases it
was a VM error that caused the failure and so the failure count was
incorrect.

This change clarifies the messaging, and prints the number of successes
even when there were some failures.

BUG=chromium:829934
TEST=run on device and inspect error messages

Change-Id: I81f327bbd281e20f9aa1beb5a527cec2e6af2408
Reviewed-on: https://chromium-review.googlesource.com/1458720
Commit-Ready: Fletcher Woodruff <fletcherw@chromium.org>
Tested-by: Fletcher Woodruff <fletcherw@chromium.org>
Reviewed-by: Stephen Barber <smbarber@chromium.org>