cros-disks: Avoid passing 'nosymfollow' for sshfs
The sshfs program chokes when passed this argument (in contrast to
passing the arg to the mount() syscall, which is supported by the
chromiumos LSM). So don't do it. We can find a way to have sshfs mount
FUSE-based filesystems with the 'nosymfollow' arg if need be. I filed
crbug.com/935712 to track that work.

BUG=chromium:925188
TEST=manually running sshfs binary with and without arg while crostini
container is running

Change-Id: I75bc7309ca304aa77c1b5daf542ccaee28441ab6
Reviewed-on: https://chromium-review.googlesource.com/1459098
Commit-Ready: Micah Morton <mortonm@chromium.org>
Tested-by: Micah Morton <mortonm@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Sergei Datsenko <dats@chromium.org>