vm_tools: sommelier: Fix pasting of data larger than 4k
Sommelier wasn't handling WL_EVENT_HANGUP events from the incoming file
descriptor for transfer operations and was stopping reading too early.
This fixes that problem, however, there is also a bug in GTK which
prevents this from always working properly as well, which a fix exists
for but only in a newer version of GTK. Details are in the associated
bug.

BUG=chromium:926982
TEST=Verified pasting of larger data sizes works

Change-Id: Ie8d5d83f580492d939c016a022814a824f6dab73
Reviewed-on: https://chromium-review.googlesource.com/1460090
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-by: Zach Reizner <zachr@chromium.org>