vm_tools: Stop concierge from having a spurious error in DestroyDiskImage
Currently vm_concierge treats the vm image directory not existing as an
error. This can cause the uninstaller to fail if the inital crostini
install did not proceed far enough to create the directory in the first
place, which in turn makes it impossible to reinstall crostini. This CL
changes a flag passed to GetDiskPathFromName that causes it to create
the parent directory if it does not exist instead of reporting an error.

Another possibility is to split out getting the path to the image from
checking if the path exists, but that would require larger changes.

BUG=chromium:930037
TEST=manually tested
Change-Id: If97e1955dcaa8e39a3b884db2c67a319c17bf9f7
Reviewed-on: https://chromium-review.googlesource.com/1460279
Commit-Ready: Fergus Dall <sidereal@google.com>
Tested-by: Fergus Dall <sidereal@google.com>
Reviewed-by: Stephen Barber <smbarber@chromium.org>