arc-setup: Correct for subshell in arc-ureadahead
arc-ureadahead attempts to exit early if its pack file is successfully
read by ureadahead.  When the script was migrated to use logger, the
commands were moved into a compound group.  Because the compound group
is piped to logger, it is run in a subshell, so the exit doesn't have
any effect on the parent shell.  This means that the fallback arc-setup
command is always run.

The commands in the group always remove the packfile if it isn't valid,
so instead of assuming that an early exit happens, we can just check for
the pack file before running arc-setup.  If it's still there, it must
have been valid; if not, we need to run the fallback.

BUG=chromium:930144
TEST=arc-setup --mode=read-ahead isn't run when pack file exists

Change-Id: Ia21c4f45f3ba0a755b50aac9ddcda7a18ef98718
Reviewed-on: https://chromium-review.googlesource.com/1460341
Commit-Ready: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>