shill: routing: Add RemoveRoute and {Add,Remove}RouteRaw methods
Prior to this change, the RoutingTable public API only allowed clients
to add routes to a table, not to remove them. Internal code to remove
routes from a table was done by manually calling ApplyRoute.

This change introduces a RemoveRoute method and adds *Raw methods
which allow for the addition/removal of routes without modifying
the internal table bookkeeping. This is particularly useful when
replacing routes or when dealing with routes that were added
outside of shill (and hence might not yet have any bookkeeping
information).

BUG=chromium:930280
TEST=-Unit tests are passing.
     -Regular ethernet/wifi connections still working as expected
      with both IPv4 and v6.

Change-Id: I2e9c1ffd3a9a945db4793198f6da8c42e6099674
Reviewed-on: https://chromium-review.googlesource.com/1460349
Commit-Ready: Alex Khouderchah <akhouderchah@chromium.org>
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Alex Khouderchah <akhouderchah@chromium.org>