shill: routing: Allow specifying an allocated table as per-device
Per-device routing tables let us employ the power of policy-based
routing in choosing which device to use for outgoing network traffic
of particular characteristics. Scenarios like "only send traffic out
of interface N if the packets have source address X" or "prefer to
send traffic with fwmark Y out of interface M" cannot be achieved
using only routes in the main routing table.

The RoutingTable code prior to this change allows for the creation of
new kernel routing tables for use in policy based routing (at least
from the perspective of the user). However, these plainly-allocated
routing tables are not suitable for use as per-device routing
tables. In particular, routing table entries that come from outside of
shill (e.g. entries added by the kernel on receipt of IPv6 router
advertisements, or for setting up default routes to a specific
gateway) still go to the main routing table.

This change allows for a table to be specified as the per-device table
for a particular interface. When this is performed, routing table
entries for that interface will be placed into the per-device table
rather than in a different table.

BUG=chromium:930280
TEST=-All unit tests are passing.
     -Modified Connection code such that all connections use per-device
      tables.
        -Was able to change which connection (ethernet, wifi, or
         cellular) was used by re-ordering the routing rules using
         ip rule and ip -6 rule. The interface used was verified by
         checking which IP address a remote server was seeing.
        -Was able to prevent outgoing network traffic from being sent
         by removing the per-device tables from the routing rules.
         That is, with only the main routing table, remote hosts were
         indicated as being unreachable despite the interfaces being
         connected.

Change-Id: I01f380cb8a16fe09095789e80a26e94f0439c70c
Reviewed-on: https://chromium-review.googlesource.com/1460351
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Alex Khouderchah <akhouderchah@chromium.org>