shill: routing: Keep track of blackhole routes
RoutingTable::AllocTableId is a method which seeks to provide the
caller with the ID of an empty routing table. In reality, the method
pulls a table ID from an MRU id list and clears any routing entries
that RoutingTable has kept track of for that particular table. Thus
RoutingTable's ability to keep track of the routing entries added to
kernel routing tables is essential for the overall RoutingTable
functionality to work as expected. In particular, the failure to keep
track of a routing entry can result in AllocTableId returning the ID
of a table that still contains some routing entries.

In the case of CreateBlackholeRoute, a blackhole route is added to a
routing table without the entry being kept track of. In this way a
table that has a blackhole route will continue to have it, even after
the table is freed and re-allocated. This change fixes this issue,
which has only not been a disaster so far because the Connection code
tends to either not use extra tables at all or to reuse the same table
id for the same purpose.

Note that there is a single difference in the RTNLMessage
sent, which is that RTA_DST is set (by ApplyRoute, which sets this
attribute unconditionally) to an empty address, whereas the old
blackhole code did not set RTA_DST at all. This does not seem to
affect the behavior, but is worth keeping in mind if aberrant
blackhole behavior shows up in the near future.

BUG=chromium:930280
TEST=-All unit tests are passing.
     -Manually verified that created blackholes still work as before,
      despite setting the RTA_DST attribute in the rtnl message.
     -Created a routing table containing a blackhole route, and ensured
      that freeing and re-allocating that table id in shill would result
      in the corresponding kernel table containing no entries (prior to
      this change, the table would still contain the blackhole route).

Change-Id: I95d98abb9955a479f3adae3f851a6c07c23674c6
Reviewed-on: https://chromium-review.googlesource.com/1460350
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>