system_api: Revert crrev.com/c/1300197
Remove the proto fields added to support a "dependency message" in
Crostini uninstall UI. This was never implemented and I don't think it
was the right approach. (See comment on UX slide deck.) If the new owner
of this feature wants to implement something like this, they should
decide how best to represent it in a proto.

This field was never used anywhere, so we shouldn't need to reserve the
tag number.

BUG=chromium:822514
TEST=Compiled

Change-Id: Icdbe7dd0a26404afdae9f461dcdf2df12d81fd7f
Reviewed-on: https://chromium-review.googlesource.com/1464143
Commit-Ready: Ian Barkley-Yeung <iby@chromium.org>
Tested-by: Ian Barkley-Yeung <iby@chromium.org>
Reviewed-by: Jeffrey Kardatzke <jkardatzke@google.com>