shill: cellular: Get exit codes from pppd
pppd has an option called max_fail. When max_fail > 0, pppd will try
to connect to the peer max_fail times before exiting with an
informative exit code. Cellular, by not setting this option, uses the
default max_fail value of 10 prior to this change. However, cellular
code also considers any disconnect from pppd to be a failure and shuts
down pppd when this occurs.

In this way, cellular essentially uses max_fail=1, but in a roundabout
way that doesn't involve pppd being aware of the fact. This CL changes
such behavior so that both shill and pppd have coherent views of how
connections should be made, allowing for cellular code to use the
error code returned from pppd in order to determine what kind of
failure, if any, occurred. As a bonus, attempting to introduce
multiple PPP connection attempts before considering the connection a
failure is now as simple as changing the max_fail value.

BUG=chromium:929781
TEST=-All unit tests are passing.
     -Connecting to a 3G dongle and then removing the dongle results
      in no failure; pppd exits with EXIT_OK.
     -Connecting to a 3G dongle and then sending SIGTERM to the
      spawned pppd process causes a service failure of type
      kFailureUnknown; pppd exits with an error (specifically
      EXIT_USER_REQUEST).

Change-Id: Ia86501561c159a1301902ab594299840581de6d9
Reviewed-on: https://chromium-review.googlesource.com/1464146
Commit-Ready: Alex Khouderchah <akhouderchah@chromium.org>
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>