camera: usb: Simplify and fix AE antibanding mode
Always use AUTO for ANDROID_CONTROL_AE_ANTIBANDING_MODE.

We don't respect ANDROID_CONTROL_AE_ANTIBANDING_MODE in the capture
request at all [1], so we might set it to 60Hz based on the current
location even if the capture request set it to 50Hz. Changing it to AUTO
matches the current behavior better.

Note that we still report AUTO even if the underlying camera device does
not support setting power line frequency. This is required for CTS [2]
and is the behavior of the external camera provider in Android [3].

[1] https://chromium.git.corp.google.com/chromiumos/platform2/+/1bb6d1de9aaf154931fa84c176f3577325714fea/camera/hal/usb/v4l2_camera_device.cc#188
[2] http://cs/android/cts/tests/camera/utils/src/android/hardware/camera2/cts/helpers/StaticMetadata.java?l=568&rcl=5bb489498191a32006744e789a7c8509613c5b00
[3] http://cs/android/hardware/interfaces/camera/device/3.4/default/ExternalCameraDevice.cpp?l=259-262&rcl=afa04bc7d04b0f86056ef813cc0e894251908079

BUG=b:122942809,b:77833131
TEST=Run CameraDeviceTest#testCameraDevicePreviewTemplate with an
     external camera that does not support setting power line frequency.

Change-Id: I3a85022a71783152e378df35dd36a67ab793da33
Reviewed-on: https://chromium-review.googlesource.com/1469087
Commit-Ready: Shik Chen <shik@chromium.org>
Tested-by: Shik Chen <shik@chromium.org>
Reviewed-by: Heng-ruey Hsu <henryhsu@google.com>