shill: Add EthernetProvider
Implement EthernetProvider without using it in Manager for ease of
debugging in the future. Currently, EthernetServices are not
created until the device comes up. Since using an EthernetProvider
will enable a few Manager dbus method calls, EthernetProvider needs
to create a general EthernetService so that it can be configured
even if the Ethernet device is not up yet.

BUG=chromium:916041
TEST=unittests

Change-Id: I1a1d0d331a5e5c167d94898b2eb31aaa40015f87
Reviewed-on: https://chromium-review.googlesource.com/1475043
Commit-Ready: Matthew Wang <matthewmwang@chromium.org>
Tested-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>