shill: fix default blackhole route creation
This CL fixes a regression in the creation of default blackhole routes
introduced by CL:1319984. The default blackhole routes should only be
created when there is some UID or IP address to blackhole. CL:1319984
relies on a nullptr check on |IPConfig::Properties::blackholed_addrs| to
determine if there is any IP address to blackhole. However, that's
incorrect because |IPConfig::Properties::blackholed_addrs| is derived
from |Device::blackholed_addrs_|, which is always not null even if there
is no IP address to blackhole.

As a minimal fix to address the regression, this CL explicitly checks if
|IPConfig::Properties::blackholed_addrs| is empty or not. A follow-up
CL should simplify the code path by avoiding nullptr for
|IPConfig::Properties::blackholed_addrs|.

BUG=chromium:933527
TEST=Tested the following:
- Run unit tests.
- Run network_VPNConnect tests.
- Test IP connectivity after connecting to various types of VPN service:
  L2TP/IPsec, OpenVPN, Chrome VPN app, Android VPN app.

Change-Id: I317bb799957d1d61d9ec1e957643ecf09cc79193
Reviewed-on: https://chromium-review.googlesource.com/1475782
Commit-Ready: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Alex Khouderchah <akhouderchah@chromium.org>