arc: networkd: Enable/disable multi-networking based on Android and Chrome version.
This patch introduces the notion of multinetworking being enabled (or not).
This is necessary in order to change the behavior of the arcbr0/arc0
interfaces depending on what it must support. In particular, in single
network legacy mode, this interface handles all inbound and outbound
traffic, where the former is only ever supported from the default Shill
service. But in multinet mode, these need only maintain the well-known
subnet mapping in order to support ADB. Further, there should be no other
traffic ingress/egress on this interface and therefore iptables config
should be different.

Two mechanisms are introduced to toggle the use of multi-networking:
the default behavior is to examine the /etc/lsb-releases data for
the ARC Android and the Chrome release versions. For ARC Android, P is
the minimum required version. For Chome, this patch disables multinet
by setting the minimum version to INT_MAX. This will be changed at the
appropriate time. A flag is also introduced to force multi-networking
on or off.

BUG=b:77293260
TEST=ran presubmit for nocturne that includes: arc-networkd, shill and
vm_host_tools unit tests, and various Tast and Autotest tests.
TEST=manual verification on flashed nocturne.

Change-Id: Ifbb8e5819937d8590c9bacde1867409f862c7f45
Reviewed-on: https://chromium-review.googlesource.com/1478533
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Garrick Evans <garrick@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>