vm_tools: p9: Update wire_format_derive's dependencies
Update the wire_format_derive crate to use the newer versions of syn,
quote, and proc_macro2.  We already did this in the crosvm repo but it
seems to have been dropped when we moved the code into platform2.

BUG=none
TEST=unit tests still pass
CQ-DEPEND=CL:1479111

Change-Id: I57efcaa56ecaba5a4f8d7629be15d77b5ef634b9
Reviewed-on: https://chromium-review.googlesource.com/1479150
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Reviewed-by: David Tolnay <dtolnay@chromium.org>